Workshop Updates (#15)
* Workshop Updates

Fix a few spelling mistakes and rework some of the CI/CD section.

DeploymentConfigs are similar to Routes in that they are an OpenShift
concept developed before the Kubernetes native concept of Deployments
were created. A distinct advantage of DeploymentConfigs is that updates
can be triggered by things like ImageStreams, but it makes some
trade offs in being less flexible when it comes to rolling out updates.

Some discussion of it here: https://docs.openshift.com/container-platform/4.1/applications/deployments/what-deployments-are.html#deployments-comparing-deploymentconfigs_what-deployments-are

* Add a source label

Adds a label of `source: caas-workshop` to make cleaning up the workshop
space easier in the future. We can just:
`oc delete x -l='source=caas-workshop'`