Add workshop cleanup scripts (#21)
This Jenkinsfile is intended to be triggered manually by someone after
the workshop has completed to clean up any artifacts created by the
participants.

It uses `oc` to delete any artifacts created in OpenShift that have the
label `source=caas-workshop` and it runs the Python script to delete all
the tags from the `registry.ford.com/devenablement/workshop` Quay
repository. The version of Python on Jenkins is 2.7 which is what the
script targets right now but I will likely make a new container image
that has Python 3 in it and migrate to that so we can stay up to date.