Updated file names, titles and some grammar/clarity
Lessons 1-9 have had their file names updated with a leading 0 for more
consistent sorting.

A few lessons had file names, links to the next lesson and titles
updated to hopefully be clearer.

Added some sections to volumes and manifests where we relied on
presenting the content and did not write down all the meaningful
information.

Removed a bunch of trailing whitespace so review with "ignore
whitespace changes" turned on.