Update URL for fetching tags
The latest update to Quay moved the `/repository/{repo}/tag` URL to
`/repository/{repo}/tag/`. That trailing `/` is necessary and the
previous endpoint returns a 308 Permanent Redirect.

To avoid using any dependencies in the Python script we are using a very
low-level HTTP client, which doesn't follow redirects by default and
threw an error when encountering that response. Implementing following
all redirects seems out of the scope of the goals of the script for now.