Fix issues (#29)
For some reason git couldn't find glibc in the replica image backing jenkins-pipline. This is the image: registry.ford.com/sde/jenkins_jnlp_slave_official_openshift:3.23-1.g, but that repo appears to be private. Since this job takes about 10 seconds to run I switched to running it on the master Jenkins node.

I also had to update the permissions on the role used to grant permissions in OpenShift. The existing role seems to have stopped working in the move from OpenShift 4.3 to 4.4, but I couldn't find any patch notes related to the change.